+++ lunes_cms/__init__.py
@@ -1,4 +1,4 @@
 """
 Content Management System for the Lunes Vocabulary Trainer App
 """
-__version__ = "2023.2.0"
+__version__ = "2023.5.0"